web/requirements.txt
kivy==2.2.1
kivymd==1.1.1
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
websockets==12.0
pillow==10.1.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
certifi==2023.11.17
urllib3==2.1.0
geopy==2.4.1
flask-socketio==5.3.6
eventlet==0.33.3
plyer==2.1.0
